ref: Refactored code to MVC model
