--- app.js
@@ -1,114 +1,0 @@
-require("dotenv").config();
-require("./config/database").connect();
-const express = require("express");
-const bcrypt = require("bcryptjs");
-const jwt = require("jsonwebtoken");
-
-const User = require("./model/user");
-const passwordReset = require("./passwordReset");
-const auth = require("./middleware/auth");
-
-const app = express();
-
-app.use(express.json({ limit: "50mb" }));
-
-app.post("/api/v1/signup", async (req, res) => {
-  try {
-    // Get user input
-    const { firstName, lastName, email, password } = req.body;
-
-    // Validate user input
-    if (!(email && password && firstName && lastName)) {
-      res.status(400).send("All input is required");
-    }
-
-    // check if user already exist
-    // Validate if user exist in our database
-    const oldUser = await User.findOne({ email });
-
-    if (oldUser) {
-      return res.status(409).send("User Already Exist. Please Login");
-    }
-
-    // Encrypt user password
-    encryptedUserPassword = await bcrypt.hash(password, 10);
-
-    // Create user in our database
-    const user = await User.create({
-      first_name: firstName,
-      last_name: lastName,
-      email: email.toLowerCase(), // sanitize: convert email to lowercase
-      password: encryptedUserPassword,
-    });
-
-    // Create token
-    const token = jwt.sign(
-      { user_id: user._id, email },
-      process.env.TOKEN_KEY,
-      {
-        expiresIn: "5h",
-      }
-    );
-    // save user token
-    user.token = token;
-
-    // return new user
-    res.status(201).json(user);
-  } catch (err) {
-    console.log(err);
-  }
-});
-
-app.post("/api/v1/login", async (req, res) => {
-  try {
-    // Get user input
-    const { email, password } = req.body;
-
-    // Validate user input
-    if (!(email && password)) {
-      res.status(400).send("All input is required");
-    }
-    // Validate if user exist in our database
-    const user = await User.findOne({ email });
-
-    if (user && (await bcrypt.compare(password, user.password))) {
-      // Create token
-      const token = jwt.sign(
-        { user_id: user._id, email },
-        process.env.TOKEN_KEY,
-        {
-          expiresIn: "5h",
-        }
-      );
-
-      // save user token
-      user.token = token;
-
-      // user
-      return res.status(200).json(user);
-    }
-    return res.status(400).send("Invalid Credentials");
-  } catch (err) {
-    console.log(err);
-  }
-});
-
-app.get("/api/v1/welcome", auth, (req, res) => {
-  res.status(200).send("Welcome to Koobo");
-});
-
-app.use("/api/v1/password-reset", passwordReset);
-
-// This should be the last route else any after it won't work
-app.use("*", (req, res) => {
-  res.status(404).json({
-    success: "false",
-    message: "Page not found",
-    error: {
-      statusCode: 404,
-      message: "You reached a route that is not defined on this server",
-    },
-  });
-});
-
-module.exports = app;
