--- config/webhook.js
@@ -1,6 +1,0 @@
-const app = require("../app");
-
-app.post("/hook", (req, res) => {
-    console.log(req.body);
-    res.status(200).json('Webhook is Active').end();
-});
