+++ index.js
@@ -1,5 +1,6 @@
-const http = require("http");
-const app = require("./app");
+const http = require('http');
+const app = require('./src/app');
+
 const server = http.createServer(app);
 
 const { API_PORT } = process.env;
