+++ package.json
@@ -24,14 +24,19 @@
     "cors": "^2.8.5",
     "dotenv": "^14.3.0",
     "express": "^4.17.2",
+    "helmet": "^5.0.2",
     "joi": "^17.6.0",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^6.1.8",
+    "morgan": "^1.10.0",
     "nodemailer": "^6.7.2"
   },
   "devDependencies": {
     "chai": "^4.3.4",
     "chai-http": "^4.3.0",
+    "eslint": "^8.11.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-plugin-import": "^2.25.4",
     "mocha": "^9.2.0",
     "nodemon": "^2.0.15",
     "supertest": "^6.2.2"
