+++ src/config/database.js
@@ -1,4 +1,5 @@
-const mongoose = require("mongoose");
+const mongoose = require('mongoose');
+
 const { MONGO_URI } = process.env;
 
 exports.connect = () => {
@@ -6,10 +7,10 @@ exports.connect = () => {
   mongoose
     .connect(MONGO_URI)
     .then(() => {
-      console.log("Successfully connected to database");
+      console.log('Successfully connected to database');
     })
     .catch((error) => {
-      console.log("database connection failed. exiting now...");
+      console.log('database connection failed. exiting now...');
       console.error(error);
       process.exit(1);
     });
