+++ src/controllers/authentication/login.controller.js
@@ -1,0 +1,40 @@
+/* eslint-disable consistent-return */
+/* eslint-disable no-underscore-dangle */
+require('dotenv').config();
+const bcrypt = require('bcryptjs');
+const jwt = require('jsonwebtoken');
+const User = require('../../model/user');
+
+exports.loginUser = async (req, res) => {
+  try {
+    // Get user input
+    const { email, password } = req.body;
+
+    // Validate user input
+    if (!(email && password)) {
+      res.status(400).send('All input is required');
+    }
+    // Validate if user exist in our database
+    const user = await User.findOne({ email });
+
+    if (user && (await bcrypt.compare(password, user.password))) {
+      // Create token
+      const token = jwt.sign(
+        { user_id: user._id, email },
+        process.env.TOKEN_KEY,
+        {
+          expiresIn: '5h',
+        },
+      );
+
+      // save user token
+      user.token = token;
+
+      // user
+      return res.status(200).json(user);
+    }
+    return res.status(400).send('Invalid Credentials');
+  } catch (err) {
+    console.log(err);
+  }
+};
