+++ src/controllers/authentication/signup.controller.js
@@ -1,0 +1,53 @@
+require('dotenv').config();
+const bcrypt = require('bcryptjs');
+const jwt = require('jsonwebtoken');
+const User = require('../../model/user');
+
+exports.signupUser = async (req, res) => {
+  try {
+    // Get user input
+    const {
+      firstName, lastName, email, password,
+    } = req.body;
+
+    // Validate user input
+    if (!(email && password && firstName && lastName)) {
+      res.status(400).send('All input is required');
+    }
+
+    // check if user already exist
+    // Validate if user exist in our database
+    const oldUser = await User.findOne({ email });
+
+    if (oldUser) {
+      return res.status(409).send('User Already Exist. Please Login');
+    }
+
+    // Encrypt user password
+    encryptedUserPassword = await bcrypt.hash(password, 10);
+
+    // Create user in our database
+    const user = await User.create({
+      first_name: firstName,
+      last_name: lastName,
+      email: email.toLowerCase(), // sanitize: convert email to lowercase
+      password: encryptedUserPassword,
+    });
+
+    // Create token
+    const token = jwt.sign(
+      { user_id: user._id, email },
+      process.env.TOKEN_KEY,
+      {
+        expiresIn: '5h',
+      },
+    );
+      // save user token
+    user.token = token;
+
+    // return new user
+    res.status(201).json(user);
+  } catch (err) {
+    console.log(err);
+  }
+};
