+++ src/controllers/users/home.controller.js
@@ -1,0 +1,3 @@
+exports.homePage = (req, res) => {
+  res.status(200).send('Welcome to Koobo');
+};
