+++ src/middleware/auth.js
@@ -1,19 +1,18 @@
-const jwt = require("jsonwebtoken");
+const jwt = require('jsonwebtoken');
 
 const config = process.env;
 
 const verifyToken = (req, res, next) => {
-  const token =
-    req.body.token || req.query.token || req.headers["x-access-token"];
+  const token = req.body.token || req.query.token || req.headers['x-access-token'];
 
   if (!token) {
-    return res.status(403).send("A token is required for authentication");
+    return res.status(403).send('A token is required for authentication');
   }
   try {
     const decoded = jwt.verify(token, config.TOKEN_KEY);
     req.user = decoded;
   } catch (err) {
-    return res.status(401).send("Invalid Token");
+    return res.status(401).send('Invalid Token');
   }
   return next();
 };
