+++ src/model/token.js
@@ -1,0 +1,22 @@
+const mongoose = require('mongoose');
+
+const { Schema } = mongoose;
+
+const tokenSchema = new Schema({
+  userId: {
+    type: Schema.Types.ObjectId,
+    required: true,
+    ref: 'user',
+  },
+  token: {
+    type: String,
+    required: true,
+  },
+  createdAt: {
+    type: Date,
+    default: Date.now,
+    expires: 3600,
+  },
+});
+
+module.exports = mongoose.model('token', tokenSchema);
